visualization/requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Django==3.2
django-cors-headers==3.11.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
fastapi==0.74.0
h11==0.13.0
idna==3.3
numpy==1.22.3
Pillow==9.0.1
pydantic==1.9.0
pytz==2021.3
requests==2.27.1
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.17.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.5
